[klibc] dash: [ERROR] Set exitstatus in onint
[ dash commit b4ce4120f87d89476b2d6ab31df43900d2f5ce89 ]

Currently the exit status when we receive SIGINT is set in evalcommand
which means that it doesn't always get set.  For example, if you press
CTRL-C at the prompt of an interactive dash, the exit status is not
set to 130 as it is in many other Bourne shells.

This patch fixes this by moving the setting of the exit status into
onint which also simplifies evalcommand.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>