[klibc] dash: [BUILTIN] Return without arguments in a trap should use status outside traps
[ dash commit 70c16dd30d4cf824aa895e9f6c095fec741c65a8 ]

POSIX now requires that return without arguments in a trap should
return the last command status prior to executing traps.  This
patch implements this behaviour.

Incidentally this also changes the behaviour of return without
arguments in a loop conditional to use the last exit status in
the body as opposed to the last command in the conditional when
there is one.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>