[klibc] dash: eval: Fix exit status when calling eval/dot with no commands
[ dash commit 17a5f24e0a8ec22f74399764db30d97ae310f3c6 ]

On 17/11/2015 03:18, Gioele Barabucci wrote:
> Hello,
>
> a bug has been filed in the Debian BTS about dash not resetting the exit
> status after sourcing an empty file with the dot command. [1]
>
> The following test echoes "OK" with bash and "fail" with dash
>
>      #!/bin/sh
>
>      echo > ./empty
>      false
>
>      . ./empty && echo "OK" || echo "fail"
 >
> A similar bug in dash has been discussed and addressed in 2011 [2], but
> it looks like the solution has been only partial.
>
> The version of dash I tested is the current git master branch, commit
> 2e58422.
>
> [1] https://bugs.debian.org/777262
> [2] http://article.gmane.org/gmane.comp.shells.dash/531

The bug described there was about empty files. While the fix has been
applied and does make dash handle empty files properly, your test
doesn't use an empty file, it uses a file containing a single blank
line. Unfortunately, the single blank line gets parsed by dash as a null
command, null commands don't (and shouldn't) reset the exit status, and
the fix you link to doesn't handle this because it sees a command has
been executed and saves the exit status after executing that command as
the exit status to be used by ".".

I think the easiest way to fix this is to prevent null commands from
affecting status in cmdloop, as attached.

An alternative could be to change the outer if condition to exclude
n == NULL, but I didn't do that because the change of job_warning and
clearing of numeof make sense to me even for null commands. Besides,
when debug tracing is enabled, null commands have a visible effect that
should remain.

Note that this fixes the problem with . but the same problem can be
present in other locations. For example,

     false
     eval "
     " && echo OK || echo Fail

used to print Fail, and needed the same modification in the evalstring
function to make that print OK (included in the attached patch). There
may be other similar bugs lurking.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>