[klibc] dash: exec: Do not allocate stack string in padvance
[ dash commit 4f7527f8e49276894751a9b100e68bc46568bb85 ]

Many callers of padvance immediately free the allocated string
so this patch moves the stalloc call to the caller.  Instead of
returning the allocated string, padvance now returns the length
to allocate (this may be longer than the actual string length,
even including the NUL).  For the case where we would previously
return NULL, we now return -1.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>