[klibc] dash: expand: Merge syntax/quotes in memtodest with flags
[ dash commit cbca3d6f744e39d812817405f1eeef72480bf89c ]

The function arguments syntax and quotes are both derived from
the expansion flags.  As syntax is only used by memtodest we do
not need to maintain it outside of the function at all.

The only place that uses something other than BASESYNTAX or DQSYNTAX
is exptilde.  However in that case DQSYNTAX has exactly the same
effect as SQSYNTAX.

This patch merges these two arguments into a single flags.  The
macro QUOTES_KEEPNUL has been renamed to EXP_KEEPNUL in order
to keep the namespace separate.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>